requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
replicate==1.0.7
python-dotenv==1.1.0
requests==2.31.0